move collator to utils
