+++ trl/trainer/utils.py
@@ -214,6 +214,64 @@ class DataCollatorForCompletionOnlyLM(DataCollatorForLanguageModeling):
         return batch
 
 
+class DataCollatorForLastCompletionLM(DataCollatorForLanguageModeling):
+    """
+    Data collator for language modeling that ignores all tokens except the last completion.
+    It also separates prompts and completions for easy access.
+
+    Args:
+        tokenizer: The tokenizer to use for encoding the data.
+        mlm (bool): Whether to use masked language modeling. Default is False.
+        response_template (str): The template that marks the start of a response.
+        ignore_index (int): The index to use for ignoring tokens in loss calculation.
+    """
+
+    def __init__(
+        self, tokenizer, mlm: bool = False, response_template: str = "### Response:\n", ignore_index: int = -100
+    ):
+        super().__init__(tokenizer=tokenizer, mlm=mlm)
+        self.response_template = self.tokenizer.encode(response_template, add_special_tokens=False)
+        self.ignore_index = ignore_index
+
+    def torch_call(self, examples: List[Dict[str, Any]]) -> Dict[str, torch.Tensor]:
+        batch = super().torch_call(examples)
+
+        prompts = []
+        completions = []
+
+        for i in range(len(examples)):
+            input_ids = batch["input_ids"][i]
+            labels = batch["labels"][i]
+
+            # Find all occurrences of the response template
+            response_starts = [
+                j
+                for j in range(len(input_ids) - len(self.response_template) + 1)
+                if input_ids[j : j + len(self.response_template)].tolist() == self.response_template
+            ]
+
+            if not response_starts:
+                # If no response template is found, treat the whole input as a prompt
+                prompts.append(input_ids)
+                completions.append(torch.tensor([]))
+                labels[:] = self.ignore_index
+            else:
+                # Get the start of the last response
+                last_response_start = response_starts[-1]
+
+                # Separate prompt and completion
+                prompts.append(input_ids[:last_response_start])
+                completions.append(input_ids[last_response_start:])
+
+                # Set labels for all tokens before the last response to ignore_index
+                labels[:last_response_start] = self.ignore_index
+
+        # Add prompts and completions to the batch
+        batch["prompts"] = pad(prompts, padding_value=self.tokenizer.pad_token_id, padding_side="left")
+        batch["completions"] = pad(completions, padding_value=self.tokenizer.pad_token_id, padding_side="right")
+
+        return batch
+
 @dataclass
 class RewardDataCollatorWithPadding:
     r"""
